+++ install.sh
@@ -79,9 +79,20 @@ set -euo pipefail
   fi
   echo "release=$release"
 
+  curl_args=(
+    -fL
+    "https://api.github.com/repos/direnv/direnv/releases/$release"
+  )
+
+  if [[ -n "${DIRENV_GITHUB_API_TOKEN:-}" ]]; then
+    # note: this doesn't actually print the token value.
+    echo "using DIRENV_GITHUB_API_TOKEN for GitHub API authentication"
+    curl_args+=(-H "Authorization: Bearer ${DIRENV_GITHUB_API_TOKEN}")
+  fi
+
   log "looking for a download URL"
   download_url=$(
-    curl -fL "https://api.github.com/repos/direnv/direnv/releases/$release" \
+    curl "${curl_args[@]}" \
     | grep browser_download_url \
     | cut -d '"' -f 4 \
     | grep "direnv.$kernel.$machine\$"