+++ internal/cmd/config.go
@@ -109,9 +109,10 @@ func LoadConfig(env Env) (config *Config, err error) {
 	exePath = strings.Replace(exePath, "\\", "/", -1)
 	config.SelfPath = exePath
 
-	if config.WorkDir, err = os.Getwd(); err != nil {
-		err = fmt.Errorf("LoadConfig() Getwd failed: %w", err)
-		return
+	var wdErr error
+	if config.WorkDir, wdErr = os.Getwd(); wdErr != nil {
+		// handled by `findEnvUp`
+		return //nolint:nilnesserr
 	}
 
 	// Default Warn Timeout